+++ src/isaac_so_arm101/robots/__init__.py
@@ -8,5 +8,5 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
-from .so_arm100 import *
-from .so_arm100_roscon import *
+from .trs_so100.so_arm100 import *
+from .trs_so100_roscon.so_arm100_roscon import *
