+++ src/isaac_so_arm101/robots/trs_so100/usds/configuration/Livestream_test_sensor.usd
