+++ src/isaac_so_arm101/robots/trs_so100/usds/configuration/base_plate_layer1-v5.tmp.usd
