+++ src/isaac_so_arm101/robots/trs_so100/usds/so_100.usd
