+++ src/isaac_so_arm101/robots/trs_so100_roscon/usds/configuration/so_arm100_physics.usd
