+++ src/isaac_so_arm101/robots/trs_so100_roscon/usds/so_arm100.usd
