+++ src/isaac_so_arm101/tasks/lift/joint_pos_env_cfg.py
@@ -20,8 +20,8 @@ from isaaclab.sim.schemas.schemas_cfg import RigidBodyPropertiesCfg
 from isaaclab.sim.spawners.from_files.from_files_cfg import UsdFileCfg
 from isaaclab.utils import configclass
 from isaaclab.utils.assets import ISAAC_NUCLEUS_DIR
-from SO_100.robots import SO_ARM100_CFG, SO_ARM100_ROSCON_CFG  # noqa: F401
-from SO_100.tasks.lift.lift_env_cfg import LiftEnvCfg
+from isaac_so_arm101.robots import SO_ARM100_CFG, SO_ARM100_ROSCON_CFG  # noqa: F401
+from isaac_so_arm101.tasks.lift.lift_env_cfg import LiftEnvCfg
 
 from isaaclab.markers.config import FRAME_MARKER_CFG  # isort: skip
