+++ src/isaac_so_arm101/tasks/lift/mdp/rewards.py
